requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
flask==2.2.3
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
pydantic==1.10.5
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
slack-bolt==1.16.2
    # via -r requirements.in
slack-sdk==3.20.0
    # via slack-bolt
typing-extensions==4.5.0
    # via pydantic
urllib3==1.26.15
    # via requests
werkzeug==2.2.3
    # via flask
